refactor: extract greeting jupiter panel
