+++ src/components/dashboard/widgets/greeting-widget/GreetingWidget.tsx
@@ -1,15 +1,10 @@
-import { useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-	BentoTileChromeContext,
-	type BentoTileChromeState,
-} from '@psycron/components/dashboard/bento-tile/BentoTile.context';
 import { LummiHero } from '@psycron/components/dashboard/lummi-hero/LummiHero';
-import { JupiterInsightsWidget } from '@psycron/components/dashboard/widgets/jupiter-insights-widget/JupiterInsightsWidget';
 import { useWeather } from '@psycron/hooks/useWeather';
 import { getDateLocale } from '@psycron/utils/date/date.utils';
 import { format } from 'date-fns';
 
+import { GreetingJupiterPanel } from './GreetingJupiterPanel';
 import {
 	GreetingCopy,
 	GreetingEyebrow,
@@ -17,11 +12,6 @@ import {
 	GreetingHeadline,
 	GreetingSubtext,
 	GreetingWidgetRoot,
-	JupiterIconBox,
-	JupiterPanel,
-	JupiterPanelFooter,
-	JupiterPanelHeader,
-	JupiterPanelIdentity,
 	WeatherMeta,
 } from './GreetingWidget.styles';
 import type { GreetingWidgetProps } from './GreetingWidget.types';
@@ -29,14 +19,6 @@ import type { GreetingWidgetProps } from './GreetingWidget.types';
 const textVariants = {
 	hidden: { opacity: 0, x: -12 },
 	visible: { opacity: 1, x: 0, transition: { duration: 0.4, ease: 'easeOut' } },
-};
-
-const iconFloatAnimate = { y: [0, -4, 0] };
-const iconFloatTransition = {
-	duration: 4,
-	ease: 'easeInOut',
-	repeat: Infinity,
-	repeatType: 'loop' as const,
 };
 
 export const GreetingWidget = ({
@@ -57,16 +39,6 @@ export const GreetingWidget = ({
 		sessionCount !== undefined && sessionCount > 0
 			? t('page.dashboard.greeting.subtext-sessions', { count: sessionCount })
 			: t('page.dashboard.greeting.subtext');
-
-	// Isolated context — JupiterInsightsWidget's WidgetLayout deposits
-	// its chrome here instead of the parent BentoTile's context.
-	const [jupChrome, setJupChrome] = useState<BentoTileChromeState>({});
-	const jupCtx = useMemo(() => ({ setChrome: setJupChrome }), []);
-
-	const hasJupHeader = Boolean(
-		jupChrome.icon || jupChrome.title || jupChrome.headerActions
-	);
-	const hasJupFooter = Boolean(jupChrome.footer || jupChrome.actions);
 
 	return (
 		<GreetingWidgetRoot
@@ -123,48 +95,7 @@ export const GreetingWidget = ({
 				</GreetingCopy>
 			</GreetingHead>
 
-			<JupiterPanel
-				data-testid='dashboard-greeting-jupiter-panel'
-				id='dashboard-greeting-jupiter-panel'
-			>
-				<BentoTileChromeContext.Provider value={jupCtx}>
-					{hasJupHeader && (
-						<JupiterPanelHeader
-							data-testid='dashboard-greeting-jupiter-header'
-							id='dashboard-greeting-jupiter-header'
-						>
-							<JupiterPanelIdentity
-								data-testid='dashboard-greeting-jupiter-identity'
-								id='dashboard-greeting-jupiter-identity'
-							>
-								{jupChrome.icon && (
-									<JupiterIconBox
-										animate={iconFloatAnimate}
-										data-testid='dashboard-greeting-jupiter-icon'
-										id='dashboard-greeting-jupiter-icon'
-										transition={iconFloatTransition}
-									>
-										{jupChrome.icon}
-									</JupiterIconBox>
-								)}
-								{jupChrome.title}
-							</JupiterPanelIdentity>
-							{jupChrome.headerActions}
-						</JupiterPanelHeader>
-					)}
-					<JupiterInsightsWidget
-						compact
-						insights={insights}
-						isLoading={isLoading}
-					/>
-					{hasJupFooter && (
-						<JupiterPanelFooter>
-							{jupChrome.footer}
-							{jupChrome.actions}
-						</JupiterPanelFooter>
-					)}
-				</BentoTileChromeContext.Provider>
-			</JupiterPanel>
+			<GreetingJupiterPanel insights={insights} isLoading={isLoading} />
 		</GreetingWidgetRoot>
 	);
 };
